[Refactor] Post 관련 파일 post 디렉토리 이동에 따른 경로 재설정
